+++ mk/mk/src/script/api/entity/entity_builder_params/nine_patch_renderer.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 
 pub struct NinePatchRendererParams {
     pub layer: Layer,
-    pub order: isize,
+    pub order: i32,
     pub color: Color,
     pub shader: Arc<Shader>,
     pub nine_patch: Arc<SpriteNinePatch>,
